Move Process, ProcessFacet, Event, Tree to Platform
