+++ src/cbc_sdk/platform/__init__.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import
 
-from cbc_sdk.platform.base import PlatformModel, PlatformQueryBase
+from cbc_sdk.platform.base import PlatformModel, PlatformQueryBase, Process, Event, Tree, ProcessFacet, AsyncProcessQuery
 from cbc_sdk.platform.devices import Device, DeviceSearchQuery
 from cbc_sdk.platform.alerts import (BaseAlert, WatchlistAlert, CBAnalyticsAlert,
                                      VMwareAlert, Workflow, WorkflowStatus)
