+++ src/tests/unit/platform/test_platform_events.py
@@ -1,11 +1,11 @@
-"""Testing Event object of cbc_sdk.enterprise_edr"""
+"""Testing Event object of cbc_sdk.platform"""
 
 import pytest
 import logging
-from cbc_sdk.enterprise_edr import Event, Process
+from cbc_sdk.platform import Event, Process
 from cbc_sdk.rest_api import CBCloudAPI
 from tests.unit.fixtures.CBCSDKMock import CBCSDKMock
-from tests.unit.fixtures.enterprise_edr.mock_events import (EVENT_SEARCH_VALIDATION_RESP,
+from tests.unit.fixtures.platform.mock_events import (EVENT_SEARCH_VALIDATION_RESP,
                                                             EVENT_SEARCH_RESP_INTERIM,
                                                             EVENT_SEARCH_RESP,
                                                             EVENT_SEARCH_RESP_PART_ONE,
