+++ src/tests/unit/platform/test_platform_query.py
@@ -3,11 +3,11 @@
 import pytest
 import logging
 import time
-from cbc_sdk.enterprise_edr import Process, Event, ProcessFacet
+from cbc_sdk.platform import Process, Event, ProcessFacet
 from cbc_sdk.rest_api import CBCloudAPI
 from cbc_sdk.errors import ApiError
 from tests.unit.fixtures.CBCSDKMock import CBCSDKMock
-from tests.unit.fixtures.enterprise_edr.mock_process import (GET_PROCESS_SUMMARY_RESP,
+from tests.unit.fixtures.platform.mock_process import (GET_PROCESS_SUMMARY_RESP,
                                                              GET_PROCESS_SUMMARY_RESP_1,
                                                              GET_PROCESS_VALIDATION_RESP,
                                                              GET_PROCESS_VALIDATION_RESP_INVALID,
